Fix 2 serious bugs in the server
1) The server used to open the data file in text mode
and read its data with ifstream::read into a buffer,
using the size of the file. This created a problem, as the server
would read invalid characters after the EOF and store them into
the buffer. The problem is fixed. Now the server opens the file
in binary mode and client adds '\r' into the array of delimiters.

2) The Worker class didn't read the end position of the processing
entity (the string, transmitted to the map function) correctly
as a result of a bug in the Worker::start() method. This is also
fixed.